+++ apps/runtime-demo/3005-runtime-host/webpack.config.js
@@ -47,6 +47,7 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
       dts: {
         tsConfigPath: path.resolve(__dirname, 'tsconfig.app.json'),
       },
+      shareStrategy: 'loaded-first',
       shared: {
         lodash: {
           singleton: true,
@@ -73,10 +74,12 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
           requiredVersion: '^18.2.0',
         },
       },
-      runtimePlugins: [path.join(__dirname, './runtimePlugin.ts')],
     }),
   );
   config.optimization.runtimeChunk = false;
+  if (!config.devServer) {
+    config.devServer = {};
+  }
   config.devServer.host = '127.0.0.1';
   config.plugins.forEach((p) => {
     if (p.constructor.name === 'ModuleFederationPlugin') {