+++ packages/bridge/bridge-react-webpack-plugin/src/utis.ts
@@ -2,6 +2,34 @@ import fs from 'node:fs';
 import path from 'node:path';
 import semver from 'semver';
 
+export const checkVersion = (version: string) => {
+  // Extract the version number starting from the first digit
+  const versionMatch = version.match(/\d.*/);
+  if (!versionMatch) return 0;
+
+  const cleanVersion = versionMatch[0];
+
+  if (semver.gte(cleanVersion, '5.0.0') && semver.lt(cleanVersion, '6.0.0')) {
+    return 5;
+  } else if (semver.gte(cleanVersion, '6.0.0')) {
+    return 6;
+  }
+
+  return 0;
+};
+
+export const findPackageJson = (startPath: string): string | null => {
+  let currentPath = startPath;
+  while (currentPath !== path.parse(currentPath).root) {
+    const packageJsonPath = path.join(currentPath, 'package.json');
+    if (fs.existsSync(packageJsonPath)) {
+      return packageJsonPath;
+    }
+    currentPath = path.dirname(currentPath);
+  }
+  return null;
+};
+
 export const getBridgeRouterAlias = (
   originalAlias: string,
 ): Record<string, string> => {
@@ -18,69 +46,63 @@ export const getBridgeRouterAlias = (
     };
   }
 
-  const hasBridgeReact = '@module-federation/bridge-react' in userDependencies;
-
   let bridgeRouterAlias = {};
-  // user install @module-federation/bridge-react package or set bridgeReactRouterDomAlias
-  if (hasBridgeReact) {
-    // user install react-router-dom package
-    const reactRouterDomVersion = userDependencies['react-router-dom'];
-    let majorVersion = 0;
-    let reactRouterDomPath = '';
+  let majorVersion = 0;
+  let reactRouterDomPath = '';
 
-    // if react-router-dom version is set, use the version in package.json
-    if (reactRouterDomVersion) {
-      majorVersion = semver.major(
-        semver.coerce(reactRouterDomVersion || '0.0.0') ?? '0.0.0',
-      );
-      reactRouterDomPath = require.resolve('react-router-dom');
-    } else {
-      // if react-router-dom version is not set, reslove react-router-dom to get the version
-      reactRouterDomPath = require.resolve('react-router-dom');
-      const packageJsonPath = path.resolve(
-        reactRouterDomPath,
-        '../../package.json',
-      );
-      const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));
-      majorVersion = parseInt(packageJson.version.split('.')[0]);
-    }
+  const reactRouterDomVersion = userDependencies['react-router-dom'];
 
-    // if react-router-dom path has set alias by user, use the originalAlias
-    reactRouterDomPath = originalAlias || reactRouterDomPath;
-
-    if (majorVersion === 5) {
-      bridgeRouterAlias = {
-        'react-router-dom$':
-          '@module-federation/bridge-react/dist/router-v5.es.js',
-      };
-      try {
-        require.resolve('react-router-dom/index.js');
-      } catch (error) {
-        // if react-router-dom/index.js cannot be resolved, set the alias to origin reactRouterDomPath
-        bridgeRouterAlias = {
-          ...bridgeRouterAlias,
-          'react-router-dom/index.js': reactRouterDomPath,
-        };
-      }
-    } else if (majorVersion === 6) {
-      bridgeRouterAlias = {
-        'react-router-dom$':
-          '@module-federation/bridge-react/dist/router-v6.es.js',
-      };
-
-      try {
-        require.resolve('react-router-dom/dist/index.js');
-      } catch (error) {
-        // if react-router-dom/dist/index.js cannot be resolved, set the alias to origin reactRouterDomPath
-        bridgeRouterAlias = {
-          ...bridgeRouterAlias,
-          'react-router-dom/dist/index.js': reactRouterDomPath,
-        };
-      }
-    } else {
-      console.warn('react-router-dom version is not supported');
-    }
+  if (originalAlias) {
+    reactRouterDomPath = originalAlias;
+  } else if (reactRouterDomVersion) {
+    majorVersion = checkVersion(reactRouterDomVersion);
+    reactRouterDomPath = require.resolve('react-router-dom');
+  } else {
+    reactRouterDomPath = require.resolve('react-router-dom');
   }
+
+  const packageJsonPath = findPackageJson(reactRouterDomPath);
+
+  if (packageJsonPath) {
+    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf-8'));
+    majorVersion = checkVersion(packageJson.version);
+  } else {
+    console.warn('Unable to find package.json for react-router-dom');
+  }
+
+  if (majorVersion === 5) {
+    bridgeRouterAlias = {
+      'react-router-dom$':
+        '@module-federation/bridge-react/dist/router-v5.es.js',
+    };
+    try {
+      require.resolve('react-router-dom/index.js');
+    } catch (error) {
+      // if react-router-dom/index.js cannot be resolved, set the alias to origin reactRouterDomPath
+      bridgeRouterAlias = {
+        ...bridgeRouterAlias,
+        'react-router-dom/index.js': reactRouterDomPath,
+      };
+    }
+  } else if (majorVersion === 6) {
+    bridgeRouterAlias = {
+      'react-router-dom$':
+        '@module-federation/bridge-react/dist/router-v6.es.js',
+    };
+
+    try {
+      require.resolve('react-router-dom/dist/index.js');
+    } catch (error) {
+      // if react-router-dom/dist/index.js cannot be resolved, set the alias to origin reactRouterDomPath
+      bridgeRouterAlias = {
+        ...bridgeRouterAlias,
+        'react-router-dom/dist/index.js': reactRouterDomPath,
+      };
+    }
+  } else {
+    console.warn('react-router-dom version is not supported');
+  }
+
   console.log(
     '<<<<<<<<<<<<< bridgeRouterAlias >>>>>>>>>>>>>',
     bridgeRouterAlias,