+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,29 @@
 # [8.1.0-canary.7](https://github.com/module-federation/core/compare/nextjs-mf-8.1.0-canary.6...nextjs-mf-8.1.0-canary.7) (2023-11-21)
 
+## 8.4.13
+
+### Patch Changes
+
+- Updated dependencies [f245bb3]
+- Updated dependencies [1d9bb77]
+  - @module-federation/enhanced@0.6.0
+  - @module-federation/sdk@0.6.0
+  - @module-federation/node@2.5.10
+  - @module-federation/runtime@0.6.0
+  - @module-federation/utilities@3.1.6
+
+## 8.4.12
+
+### Patch Changes
+
+- Updated dependencies [24ba96e]
+- Updated dependencies [b90fa7d]
+  - @module-federation/runtime@0.5.2
+  - @module-federation/enhanced@0.5.2
+  - @module-federation/sdk@0.5.2
+  - @module-federation/node@2.5.9
+  - @module-federation/utilities@3.1.5
+
 ## 8.4.11
 
 ### Patch Changes